Replace errWrapf with errf
This replaces all usages of errWrapf with the new errf function. With
this, error messages read from left to right again,

    errWrapf(reason, "something went wrong")
    errf("something went wrong", reason)

Note that `errWrapf` had an unused feature: If `reason` was nil, the
returned error would be nil. I verified that there's an `err != nil`
check in all spots where we call `errWrapf` as part of this switch.